refactor(file): move misc file checks into storage directory
Also added test before making SuffixFrom shorter.
