+++ src/services/NotionService/BlockHandler/BlockHandler.ts
@@ -12,7 +12,7 @@ import axios from 'axios';
 import getDeckName from '../../../lib/anki/getDeckname';
 import sanitizeTags from '../../../lib/anki/sanitizeTags';
 import { sendError } from '../../../lib/error/sendError';
-import { S3FileName, SuffixFrom } from '../../../lib/misc/file';
+import { S3FileName, SuffixFrom } from '../../../lib/storage/file';
 import getUniqueFileName from '../../../lib/misc/getUniqueFileName';
 import Deck from '../../../lib/parser/Deck';
 import Note from '../../../lib/parser/Note';
